% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volplot.R
\name{volplot}
\alias{volplot}
\title{Calculating the Number of Business Days}
\usage{
volplot(
  FC,
  pvalue,
  t_FC = 1,
  t_pvalue = 0.05,
  mainname = "Volcano plot",
  xlabel = "log2(Fold Change)",
  ylabel = "-log10(p-value)",
  upcol = "red",
  downcol = "blue",
  tline = T
)
}
\arguments{
\item{FC}{value of log2(fold change)}

\item{pvalue}{p-value}

\item{t_FC}{Threshold of log2(fold change)}

\item{t_pvalue}{Threshold of p-value}

\item{mainname}{title of graph}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{upcol}{color of up-regulated genes (genes >= t-FC)}

\item{downcol}{color of down-regulated genes (genes <= t-FC)}

\item{tline}{line of threshold (If not necessary = F)}
}
\description{
Calculating the Number of Business Days
}
